add decorate() util
