+++ src/lib/logger/pino-logger.js
@@ -10,9 +10,3 @@ global.logger = logger
 
 export default logger
 
-export const ExceptionCategory = {
-  Human: 'human',
-  Blockchain: 'blockchain',
-  Network: 'network',
-  Unexpected: 'unexpected',
-}
