--- src/lib/utils/generalError.js
@@ -1,1 +1,0 @@
-export default errorCode => `Sorry, Some unexpected error (E${errorCode}) happened, please try again`
